Do not allow an empty string as the 'class' parameter to lash_client_open().
This commit reverts a part of fd70f809738e07cd6286880bd6b3faf1ed07ef66.
The reverted part in lash_client_open() had this note:

  "seq24 usually supplies no class name and we should be able to handle this
  situation anyway, to support liblash-less clients"

This is incorrect. Liblash-less clients are by definition clients that don't
call lash_client_open() and therefore don't need the ability to pass an empty
class string to it. Seq24 is simply buggy and needs to behave better.

The rest of the code in commit fd70f809738e07cd6286880bd6b3faf1ed07ef66 is
fine and is functionally unrelated to the part now reverted.