Remove the "data path" concept entirely from LASH.
This change means that LASH will no longer provide a directory for clients
to save files in. Instead, binary data should be stored and retrieved using
lash_write() and lash_read().

Clients are not prohibited from writing data to disk, either, but they are
required to be able to recall the data later on. This implies that most
clients that write to disk need to store and recall the data path, making
the use of lash_write() and lash_read() inevitable.

This commit leaves the now-defunct PathChange code commented out because
of the still useful bits there.